consolidated bin creation
